terminado la parte frontend de la aplicacion, implementacion de nuevos componentes y funcionalidades, nueva pagina de contrataciones con rutas, controlador, factory y seeeder

+++ app/Http/Controllers/Admin/PostController.php
@@ -10,35 +10,23 @@ use Illuminate\Support\Facades\Storage;
 
 class PostController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+    
     public function index()
     {
         return view('admin.posts.index');
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+  
     public function create()
     {
         $categorias = Category::pluck('name', 'id');
        return view('admin.posts.create', compact('categorias'));
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
+   
     public function store(Request $request)
     {
+
         $post = Post::create($request->all());
 
         if ($request->file('file')) {
@@ -51,48 +39,49 @@ class PostController extends Controller
         return redirect()->route('admin.posts.edit', compact('post'));
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+  
     public function show(Post $post)
     {
         //
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function edit(Post $post)
     {
-        return view('admin.posts.edit');
+        $this->authorize('author',  $post);
+        $categorias = Category::pluck('name', 'id');
+
+        return view('admin.posts.edit', compact('post', 'categorias'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, $post)
+   
+    public function update(Request $request, Post $post)
     {
-        //
+        $this->authorize('author',  $post);
+
+        $post->update($request->all());
+
+        if ($request->file('files')) {
+            $url = Storage::put('posts', $request->file('file'));
+       
+
+            if ($post->image) {
+                Storage::delete($post->image->url);
+
+                $post->image->update([
+                    'url' => $url]);
+            }else {
+                $post->image()->create([
+                    'url' => $url]);
+            }
+        }
+
+        return redirect()->route('admin.posts.index', $post)->with('info', 'publicacion actualizada con exito');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+ 
     public function destroy(Post $post)
     {
+        $this->authorize('author',  $post);
         $post->delete();
         return redirect()->route('admin.posts.index')->with('info', 'entrada eliminada con exito');
     }
